+++ CHANGELOG.md
@@ -1,8 +1,8 @@
 # Changelog
 
 ## Version 0.21.1 - 2021-07-30
-### Fix
-- Fixed IPresentation.SaveAs(). It did not release underling resources in the right way.
+### Fixed
+- Fixed `IPresentation.SaveAs()`. It did not release underlying resources in the right way.
 
 ## Version 0.21.0 - 2021-06-23
 ### Added